rename potato probes to tuberlens
