+++ MANIFEST.in
@@ -1,7 +1,7 @@
 include README.md
 include LICENSE
 include pyproject.toml
-recursive-include src/potato/py.typed
+recursive-include src/tuberlens/py.typed
 global-exclude __pycache__
 global-exclude *.py[co]
 global-exclude .DS_Store
